+++ typescript/infra/config/environments/mainnet3/agent.ts
@@ -409,7 +409,7 @@ const hyperlane: RootAgentConfig = {
     rpcConsensusType: RpcConsensusType.Fallback,
     docker: {
       repo,
-      tag: '6f414af-20241004-131949',
+      tag: '70bb863-20241009-193832',
     },
     gasPaymentEnforcement: gasPaymentEnforcement,
     metricAppContexts,
@@ -418,7 +418,7 @@ const hyperlane: RootAgentConfig = {
   validators: {
     docker: {
       repo,
-      tag: '6f414af-20241004-131949',
+      tag: '70bb863-20241009-193832',
     },
     rpcConsensusType: RpcConsensusType.Quorum,
     chains: validatorChainConfig(Contexts.Hyperlane),
@@ -428,7 +428,7 @@ const hyperlane: RootAgentConfig = {
     rpcConsensusType: RpcConsensusType.Fallback,
     docker: {
       repo,
-      tag: '3254472-20241009-111928',
+      tag: '8001bbb-20241010-143553',
     },
     resources: scraperResources,
   },