+++ typescript/infra/config/environments/testnet4/agent.ts
@@ -43,6 +43,7 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
   [Role.Validator]: {
     alfajores: true,
     arbitrumsepolia: true,
+    arcadiatestnet: true,
     basesepolia: true,
     berabartio: true,
     bsctestnet: true,
@@ -55,6 +56,7 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     fuji: true,
     holesky: true,
     // hyperliquidevmtestnet: false,
+    odysseytestnet: true,
     optimismsepolia: true,
     // Disabling plumetestnet on Sept 16, 2024: chain is paused for "airplane mode"
     // plumetestnet: true,
@@ -63,12 +65,15 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     sepolia: true,
     solanatestnet: false,
     soneiumtestnet: true,
+    sonictestnet: true,
     suavetoliman: true,
     superpositiontestnet: true,
+    unichaintestnet: true,
   },
   [Role.Relayer]: {
     alfajores: true,
     arbitrumsepolia: true,
+    arcadiatestnet: true,
     basesepolia: true,
     berabartio: true,
     bsctestnet: true,
@@ -81,6 +86,7 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     fuji: true,
     holesky: true,
     // hyperliquidevmtestnet: false,
+    odysseytestnet: true,
     optimismsepolia: true,
     // Disabling plumetestnet on Sept 16, 2024: chain is paused for "airplane mode"
     // plumetestnet: true,
@@ -89,12 +95,15 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     sepolia: true,
     solanatestnet: false,
     soneiumtestnet: true,
+    sonictestnet: true,
     suavetoliman: true,
     superpositiontestnet: true,
+    unichaintestnet: true,
   },
   [Role.Scraper]: {
     alfajores: true,
     arbitrumsepolia: true,
+    arcadiatestnet: true,
     basesepolia: true,
     berabartio: true,
     bsctestnet: true,
@@ -108,6 +117,7 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     fuji: true,
     holesky: true,
     // hyperliquidevmtestnet: false,
+    odysseytestnet: true,
     optimismsepolia: true,
     // Disabling plumetestnet on Sept 16, 2024: chain is paused for "airplane mode"
     // plumetestnet: true,
@@ -117,8 +127,10 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     // Cannot scrape non-EVM chains
     solanatestnet: false,
     soneiumtestnet: true,
+    sonictestnet: true,
     suavetoliman: true,
     superpositiontestnet: false,
+    unichaintestnet: true,
   },
 };
 
@@ -197,7 +209,7 @@ const hyperlane: RootAgentConfig = {
     rpcConsensusType: RpcConsensusType.Fallback,
     docker: {
       repo,
-      tag: '6c45a05-20240927-172800',
+      tag: '463b35b-20241011-161150',
     },
     blacklist: [...releaseCandidateHelloworldMatchingList, ...relayBlacklist],
     gasPaymentEnforcement,
@@ -219,7 +231,7 @@ const hyperlane: RootAgentConfig = {
     rpcConsensusType: RpcConsensusType.Fallback,
     docker: {
       repo,
-      tag: '6c45a05-20240927-172800',
+      tag: '463b35b-20241011-161150',
     },
     chains: validatorChainConfig(Contexts.Hyperlane),
     resources: validatorResources,
@@ -228,7 +240,7 @@ const hyperlane: RootAgentConfig = {
     rpcConsensusType: RpcConsensusType.Fallback,
     docker: {
       repo,
-      tag: '6c45a05-20240927-172800',
+      tag: '463b35b-20241011-161150',
     },
     resources: scraperResources,
   },