+++ typescript/infra/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @hyperlane-xyz/infra
 
+## 5.5.0
+
+### Patch Changes
+
+- Updated dependencies [2afc484a2]
+- Updated dependencies [2afc484a2]
+- Updated dependencies [3254472e0]
+- Updated dependencies [fcfe91113]
+- Updated dependencies [6176c9861]
+  - @hyperlane-xyz/sdk@5.5.0
+  - @hyperlane-xyz/utils@5.5.0
+  - @hyperlane-xyz/helloworld@5.5.0
+
 ## 5.4.0
 
 ### Patch Changes